backend/requirements.txt
# ============================================================
# RETURNKART.IN — PYTHON DEPENDENCIES
# ============================================================

# Web framework
fastapi==0.115.0
uvicorn[standard]==0.30.0

# Environment
python-dotenv==1.0.1

# Supabase
supabase==2.10.0

# Google AI
google-generativeai==0.8.3

# Google OAuth / Gmail API
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.151.0

# Data validation
pydantic==2.9.2

# HTTP
httpx==0.27.2

# Utilities
python-multipart==0.0.12
